tazwikiss/wkp_Admin: shrink to 4K, add section separators

+++ tazwikiss/rootfs/var/www/wiki/plugins/wkp_Admin.sh
@@ -11,28 +11,33 @@ admin_enable()
 	done
 }
 
+admin_download()
+{
+	cat - $1 <<EOT
+Content-Type: application/octet-stream
+Content-Length: $(stat -c %s $1)
+Content-Disposition: attachment; filename=${2:-$1}
+
+EOT
+}
+
 action()
 {
 	case "$1" in
 	list|config|admin);;
-	backup)	if [ -z "$(FILE file tmpname)" ]; then
+	backup)	file=$(FILE file tmpname)
+		if [ -z "$file" ]; then
 			file=$(mktemp -p /tmp)
 			find */ | cpio -o -H newc | gzip -9 > $file
-			cat - $file <<EOT
-Content-Type: application/octet-stream
-Content-Length: $(stat -c %s $file)
-Content-Disposition: attachment; filename=wiki-$(date '+%Y%m%d%H%M').cpio.gz
-
-EOT
+			admin_download $file wiki-$(date '+%Y%m%d%H%M').cpio.gz
 			rm -f $file
 			exit 0
 		else
-			file=$(FILE file tmpname)
 			zcat $file | cpio -idmu $(echo */ | sed 's|/||g')
 			rm -rf $(dirname $file)
 			return 1
 		fi ;;
-	*) return 1 ;;
+	*)	return 1 ;;
 	esac
 	PAGE_TITLE_link=false
 	editable=false
@@ -50,31 +55,26 @@ EOT
 		[ -n "$curpassword" ] && echo $curpassword | md5sum > $secret
 	fi
 	if [ -n "$(POST save)" ]; then
-		cat <<EOT
-Content-Type: application/octet-stream
-Content-Length: $(stat -c %s $(POST file))
-Content-Disposition: attachment; filename=$(POST file)
-
-EOT
-		cat $(POST file)
+		admin_download $(POST file)
 		exit 0
 	fi
 	[ -n "$(POST restore)" ] && mv -f $(FILE data tmpname) $(POST file)
 	admin_enable locales config- config_ ./
 	admin_enable plugins wkp_ wkp_ plugins/
+	disabled="disabled=disabled"
+	[ -n "$curpassword" ] && disabled="" && 
+	curpassword="<input type=\"hidden\" name=\"curpassword\" value=\"$curpassword\" />
+"
+	hr="$curpassword<tr><td colspan=2><hr /></td><tr />"
 	CONTENT="
 <table width=\"100%\">
 <form method=\"post\" action=\"?action=admin\">
 <tr><td><h2>$MDP</h2></td>
-<td><input type=\"text\" name=\"password\" />
-<input type=\"hidden\" name=\"curpassword\" value=\"$curpassword\" />
+<td><input type=\"text\" name=\"password\" />$curpassword
 <input type=\"submit\" value=\"$DONE_BUTTON\" name=\"setpassword\" /></td></tr>
-</form>"
-	disabled=
-	[ -z "$curpassword" ] && disabled="disabled=disabled"
-	CONTENT="$CONTENT
+</form>
 <form method=\"post\" enctype=\"multipart/form-data\" action=\"?action=admin\">
-<input type=\"hidden\" name=\"curpassword\" value=\"$curpassword\" />
+$hr
 <tr><td><h2>Plugins</h2></td>
 <td><input type=\"submit\" $disabled value=\"$DONE_BUTTON\" name=\"plugins\" /></td></tr>
 "
@@ -93,16 +93,14 @@ $plugin</b></td><td><i>$description</i></td></tr>"
 	CONTENT="$CONTENT
 </form>
 <form method=\"post\" enctype=\"multipart/form-data\" action=\"?action=admin\">
-<input type=\"hidden\" name=\"curpassword\" value=\"$curpassword\" />
-<p></p>
+$hr
 <tr><td><h2>Locales</h2></td>
 <td><input type=\"submit\" $disabled value=\"$DONE_BUTTON\" name=\"locales\" /></td></tr>
 "
 	for i in config-*.sh ; do
 		j=${i#config-}
 		j=${j%.sh}
-		[ -n "$j" ] || continue
-	CONTENT="$CONTENT
+		[ -n "$j" ] && CONTENT="$CONTENT
 <tr><td><b>
 <input type=checkbox $disabled $([ -x $i ] && echo 'checked=checked ') name=\"config_$j\" />
 $j</b></td><td><i>$(. ./$i ; echo $WIKI_TITLE)</i></td></tr>
@@ -111,10 +109,9 @@ $j</b></td><td><i>$(. ./$i ; echo $WIKI_TITLE)</i></td></tr>
 	CONTENT="$CONTENT
 </form>
 <form method=\"post\" enctype=\"multipart/form-data\" action=\"?action=admin\">
-<input type=\"hidden\" name=\"curpassword\" value=\"$curpassword\" />
-<tr></tr>
-<tr><td><h2>Configuration</h2></td><td>
-<select name="file" $disabled>
+$hr
+<tr><td><h2>Configuration</h2></td>
+<td><select name="file" $disabled>
 $(for i in template.html style.css config*.sh; do
   [ -x $i ] && echo "<option>$i</option>"; done)
 </select>
@@ -122,16 +119,14 @@ $(for i in template.html style.css config*.sh; do
 <input type=\"file\" $disabled name=\"data\" />
 <input type=\"submit\" $disabled value=\"$RESTORE\" name=\"restore\" /></td></tr>
 </form>
-<tr></tr>
 <form method=\"post\" enctype=\"multipart/form-data\" action=\"?action=backup\">
-<input type=\"hidden\" name=\"curpassword\" value=\"$curpassword\" />
+$hr
 <tr><td><h2>Data</h2></td>
 <td><input type=\"submit\" $disabled name=\"save\" value=\"$DONE_BUTTON\" />
 <input type=\"file\" $disabled name=\"file\" value=\"file\" />
-<input type=\"submit\" $disabled name=\"restore\" value=\"$RESTORE\" />"
-	CONTENT="$CONTENT
+<input type=\"submit\" $disabled name=\"restore\" value=\"$RESTORE\" />
 </td></tr>
-$(du -hs */ | awk '{ printf "<tr><td><b>%s</b></td><td><i>%s</i></td></tr>\n",$1,$2 }')
+$(du -hs */ | sed 's|\(.*\)\t\(.*\)|<tr><td><b>\1</b></td><td><i>\2</i></td></tr>|')
 </form>
 </table>
 "
